AHLRAP-1070 Ensure that PyDev plugins are not Jar'd
The Pydev build relies on being able to access the contents
of some of the plugins (as the python interpreter needs access to .py
files which bridge PyDev to IPython, for example). By defaul Tycho
doesn't respect the PDE builds feature.xml unpack request, so make it
explicit in the bundles.
See also: https://issues.sonatype.org/browse/TYCHO-555

Change-Id: Id184ad56cb4f9337b7a33e378bf8c25898180bde