AHLRAP-1371 AHLRAP-1366 Fixes to F1 Console submission
This is a squash of 5 commits.

This is a fix for AHLRAP-1373 'If you change the console too often, F1
stops working'. The problem was that the exec line in shell module would
raise an exception if an appropriate console was not available for
command submission. Once the module had the raised exception, I guess
eclipse doesn't reinitialise it again so one needs to restart PyDev to
get F1 functionality back.

I fixed this by merely using a try-except block to catch any exceptions.

AHLRAP-1371 AHLRAP-1366 Fixes to F1 Console submission

This patch carries a few changes:

- Fix for AHLRAP-1371: When the Scripting Console was visible, the code
submission would send the code to a view only window and this resulted
in an unhandled exception. This exception broke the script and it
wouldn't work again. The change I have made fixes this by handling the
exception; but also raising a console window so the code can be
submitted.

- Fix for AHLRAP-1366: When stepping through the code using F1, when the
user reaches the end of the file, we don't want to wrap around to the
beginning. It makes no sense to jump all the way back to the import
statements once the module is done running. This fix was done by
changing how the goto_next_line functions works. It now returns a
boolean for whether or not is actually moved to the next line.

- Changing some function names to use pythonic_naming_style rather than
javaNamingStyle.

- Some basic dead code removal.

- There is a lot of code that could be tidied up here, but I have
refrained from going carte blanche and renaming all functions which are
not called externally to have a _prefix; or removing comments and
replacing them with extensive docstrings. This module could use some
love.

Cosmetic changes.

- Renamed variables to reflect their privacy.

- Added some meagre documentation.

Removed re-raise which was incorrectly commited previously.

Protect from F1 submission when no Console is set.

James Blackburn found that when the Scripting Console is set
and a file is opened, and F1 is hit, an exception was thrown as
no Console was found. Unfortunately, the IConsoleView doesn't
appear to provide a way to find anything but the top of the
IConsole list; so I've just protected it now.

Also, to aid readability, I've inverted some if statements to
return and reduced the nesting level for _get_console_parts and
_show_console.

AHLRAP-1366 - Fixed the F1 loop around

This is a subsequent fix to the previous commit now using the
last_line calculated from the document.

AHLRAP-1366 - Do not iterate through file when no Console is available.

Change-Id: Ic190b8b51110fa764dbbcd7f32f4d9fc006e6f15