AHLRAP-1467 Fix IPython console completion pop-up on all characters
The completion calculator caches the characters which trigger a
completion ('.', or [a-zA-Z]) based on the user preference, and whether
there is more than one completion processor registered.

The console calls the API with 0, as the number of completion
processors. The editor calls it with the number of processors specified.
 The result is inconsistent caching depending on which is called first.
 
This makes the cache behave consistently for the specified preference. 
It's possible that there are SimpleAssistProcessors that would no longer
be invoked for every character, but, AFAICS this would only affects the
template processor, and that seems to work correctly after this change.

Change-Id: Ie6ab792fbef093b9396ef84b9d7a6a07b8edecc5