AHLRAP-1492 First cut of tab completion in the IPython console
Main entrance point in the ScriptConsoleViewer.  Completions flow
through to the pydev console backend.  For tab-completion, we use the
pure ipython backend. 

The magic 'cd' is handled specially, though perhaps in the future, the
completions interface can be beefed up such that eclipse is aware which
part of the string ipython has used to make the completion.
Change-Id: Iccaadcea0509413c9ff5fffcee622fdc14aabcda